requirements.txt
yfinance
pandas
mysql-connector-python
prometheus-client
flask